s3s: ops: signature
